+++ src/PowerModelsONM.jl
@@ -43,6 +43,9 @@ module PowerModelsONM
     # Hardware statistics
     import Hwloc
 
+    # Generate samples for robust partitions
+    import StatsBase as SB
+
     # Network Graphs
     import Graphs
     import EzXML
@@ -101,6 +104,7 @@ module PowerModelsONM
     include("prob/opf.jl")
     include("prob/mld_traditional.jl")
     include("prob/mld_block.jl")
+    include("prob/mld_block_robust.jl")
     include("prob/stability.jl")
     include("prob/switch.jl")